akimpersonate for kerberos-less testing
Use the akimpersonate feature of aklog for testing without requiring external
kerberos realm. The intent is to make testing setup significantly simplier,
since this obviates the need to install a kerberos kdc, setup a realm, and
create the set of test keytabs.

A test service key is automatically generated and written to an MIT krb5 format
keytab file for the servers and aklog.

To use this feature, set AFS_AKIMPERSONATE to True.

Note: The akimpersonate features on the OpenAFS 1.6.x branch have not pulled up
to the master branch at this time. For the time being, if testing master, point
the variable ${AKLOG} to a recent 1.6.x version of aklog.