newcell: refactor client-side csdb setup
Remove the large and confusing keyword to setup the client side
CellServDB and ThisCell files and replace with some logic in the OpenAFS
library and some new simple keywords.

Move the code to fixup the symlinks created by the bosserver the first
time it is started from the robot file to the library.

Create the client etc directory before setting the client CSDB and ThisCell
file.

When appending the CellServDB.dist, create a temp file as a regular
user, then use a simple sudo cp to install it, instead of doing a cat
within sudo. This will simplify the sudo permission checks.