import test names from the Alra test suite
Import names and test suite organization from the Arla test suite
(openafs:src/tests). Tag the tests to be implemented as 'todo'.  The original
Arla test script/program is indicated as a comment on the test case tag
metadata line.

Remove the prefix numbers on the files already created, since renaming files to
change the order is unwieldly, and besides the tests are independent and may
not depend on order.