restructure library modules
Restructure the library modules to improve the code base and make the
library easier to extend as needed.

The new layout borrows heavily from the more popular third-party RF
libraries.  The library is now called OpenAFSLibrary, and the "kerberos"
library has been changed to the more accurately named keytab keywords
group.  A of the helper modules have been moved to the util sub-package.

The lib2html tool has been added to a new document folder to generate
the library documentation.