nuke installation keywords
Remove the installation keywords and RF implementation of the new cell setup.
The install and setup OpenAFS should be done before running the test harness,
instead of try trying to do that as part of the suite setup and teardown, which
proved to be impractical.

The afsutil package provides a python implementation for installing and setting
up a working cell.