afs-robotest: aklog workaround check
Unfortunately, the akimpersonate feature of aklog is only available in
versions 1.6.10 and greater, and not on any master branch versions.  To
workaround this, we setup a 1.6.10+ version of aklog and then specify
the path to in the config. However this workaround is easy to get wrong
and easy to forget when setting up a new machine, so a special check is
added before running setup and run to verify a usable aklog is ready and
print some error messages if not.