oafs: workaround HACK for "stale" mountpoint issue
If any test that creates a new volume and mountpoint (e.g. behavior)
is run more than once without an intervening afs-robotest teardown,
the second run will fail at fs setacl with "No such device" ENODEV.

This is caused by the CM still having reference to the old volumeid from
the previous run (either in vcache or in vcache->mvid).  Hence it issues
RXAFS_FetchACL for the wrong volume/fid.  Since the volume no longer
exists (was torn down on the first run), the fileserver (after a long
delay while afs_Analyze does its 7 retry busy-ladder) replies with
VNOVOL.  This ends up being reflected back to the CM (by afs_CheckCode)
as ENODEV.

Instead of trying to figure out where we are going wrong in the CM
for this very unusual use case, just workaround the problem for now
by adding an fs checkvol to robotest create_volume, remove_volume,
and release_volume.  It is actually recommended that fs checkvol be
issued after a vos release that changes the number of replicas, so
this is not unreasonable.