afsutil: package subcommand
Add a the afsutil package subcommand to build rpm packages from the
source tree, build options do not apply when doing the package builds,
instead of abusing the afsutil build subcommand.  This make easier to
see how to build the rpms from the usage messages, as well as allowing
for packaging specific options in the future.