afsutil: avoid git clean of the wrong directory
Add some santity checks to the afsutil build command to avoid running in
a directory that does not seem to be the OpenAFS top level directory.

Add a check for a new git config option to enable the git clean before
building.  To enable the git clean, run this once in your OpenAFS git
repo;

    git config --local afsutil.clean true

Sorry for nuking your config file Joe.

[backport from commit c5c761ca7ecadadc3e9f2038884604317ec7eca9]