afsutil: introduce installer base class
Split the transarc-style installer into several classes.  Introduce a
base-class for common pre/post install and pre/post remove tasks. This
is used to setup our cell configuration and so forth.

Move the linux and solaris specific functions to respective classes and
configure which one to use before the installation process starts.

Combine the TransarcInstaller and TransarcUninstaller into a single
class and defer the dest dir path checks in the __init__ function to the
pre-installation.

Instead of separate pairs of install and remove functions for clients
and servers, change the interface to use a single pair of install and
remove functions. This simplifies the handling of installation and
removal of common components, as well as simplifying the class
interface.