setup: relax aklog work-around check
Stable releases starting with 1.6.10 have an aklog which supports
akimpersonate, so the work-around check triggers a false error when
installing/testing those versions. That is, it is not neccessary to
setup a separate aklog to use akimpersonate when testing current openafs
stable brach releases.

Just print a warning and continue if akimpersonate is enabled and the
aklog variable is not set. If the variable is set, continue to
check the version.

We really need to fix aklog.