afsutil: cell host ip address sanity checks
Many modern linux distro put a loopback address/hostname entry in the
/etc/hosts file.  Avoid using this loopback address with a message
to check your /etc/hosts file.

When trying to detect the system's ip address from the network
interfaces, do not pick a random address if there is more than one.
Instead, exit with a message saying the hosts must be specified.