afsrobot: simplify config loading
Originally, afs-robotest would create an in-memory configuration filled with
default values. These defaults could be overridden by values read from a config
file.  If the config file name was the current default file, and was not
present, the in-memory defaults would be used for commands. If the user
specified the config file, and it was not found, the script would exit with an
error message.

In practice, we always use a config file and run 'config init' at least once to
create the initial file.

So from now on, require 'config init' to be the first command to be run. This
will create a config file filled with default values.  All commands, except
'config init', will load the config from a config file.  Config init will not
try to load a config file, but rather create new files (and only new files)
filled with default values.  The in-memory configuration used during command
processing will have options loaded from the file and only from the file.

The name of the config file can be given on the command line with the
--config/-c option. The default name is ~/.afsrobotestrc/afs-robotest.conf,
which can be overridden by the env var AFS_ROBOTEST_CONF.

The 'config init' will create the initial config file filled with default
values. This is the only command that does not expect a pre-existing file.

Move the env vars back to the script where they belong.

Add new functions to load and save the config and add unit tests.