afsrobot: remove AFS_ROBOTEST_SSH env var
Instead of an extra env var, require the config option to be set.  Change the
ssh create sub-command to automatically set the keyfile options on success.
Add the --keyfile option to ssh create so a non-default name can be specified
by the user.  An existing key file can be used, in which case, the config set
command can be used to set the name in the config.