afsrobot: remove AFS_ROBOTEST_ETC env var
Remove this extra env var, since AFS_ROBOTEST_CONF includes the path, we don't
need this one.

Do not use AFS_ROBOTEST_ROOT as a default for path:root in the config file. Use
the env vars for command line options defaults only, not for config file
defaults.  That leads down to a slipper slope of complexity.