afsutil: try to lookup the cell host address
If the cell hosts are not specified, we assume the localhost is the
single cell host.  First try to find a non-loopback address from the
name to address lookup. If that fails, then try to look for a single
non-loopback address from the system's network interfaces. If there is
more than one network address, then we have to give up, since we don't
know which ones might be private.

Refactor: split _detect_cellhosts into 2 functions, _lookup_cellhosts
and _detect_cellhosts.