afsutil: support modern linux kernel module naming with transarc
Install the modern style openafs.ko kernel module to the standard linux
location when installing transarc-style binaries and the openafs.ko is
available. Run depmod so it can be loaded with `modprobe openafs`.

The openafs.ko is built when the --with-linux-kernel-packaging configure
option is given.

The init script provided with the transarc-style distribution must load
the kernel module with modprobe.  A patch is provided.  It can be applied
with:

    cat patch/*.patch | patch -d <topdir> -p1

where <topdir> is the top level directory of the openafs source tree.