afsutil: import init scripts for transarc-style installs
afsutil already provides an init script to start the bosserver and the solaris
init script from the OpenAFS distribution is modified during the installation.
Changes to the linux init script will be needed too.  Also, we will want a way
to control how the client side options are set.  So, instead of relying on the
distribution init scripts for transarc-style installs, provide our own, based
on the examples from OpenAFS.

Update the afsutil package to include the init scripts as package data files.