afsutil: use afsutil init scripts for transarc installs
Before, we would use the solaris init script provided in the distribution
files, however we want to separate the server and client restart, so that
script needs to be changed.

Instead of trying to modify the init script for solaris on the fly, which could
fail when the upstream version changes, just provide our own scripts, for
solaris and for linux, which can be updated as needed.