afsutil: reload openafs.ko too
Support reloading of the kernel module when it was built the configure
option --with-linux-kernel-packaging.  This creates a module with modern
naming and loading conventions.

Since we already to use the init script to load the solaris driver,
always run the init script to restart the client on all platforms. This
makes things more uniform and will make it easier to configure afsd
startup options.