setup: fix newcell when using actual keytabs
Fix the afsutil newcell command options to work without akimpersonate.  afsutil
needs to get an admin user token to complete the top level directories setup,
since we can't do that without a client (yet). Instead of assuming
akimpersonate is always in use, give it the same options as the other commands
(which work with or without akimpersonate.)

Make separate options for akimpersonate fake keytabs and real keytabs to avoid
clobbering the real keytab when akimpersonate is changed from yes to no. Fall
back to the old name (keytab) for compatibility sake.  The options are now:

   [kerberos]
   akimpersonate = yes | no  yes: make a fake service key
   fake_keytab = path to the fake service keytab to be generated
   afs_keytab = path to a krb5 service keytab
   user_keytab = path to a krb5 keytab with user keys for the regular
                 and admin user

The afs_keytab and user_keytab should have been created with kadmin (or
kadmin.local) before running afs-robotest setup when akimpersonate is set to
'no'.