afsrobot: fix handling of non-dynroot afsd
Commit 2b28f01 changed the default afsd options from -dynroot to none, so the
options are explicit in the config file.

Fix the handling of non-dynroot setups so the cache manager is not started
until the root volumes are online to avoid the cm from hanging in the
non-dynroot environment.

Instead of trying to detect if dynroot is enabled, which triggers false
information in the presence of afs bugs, make the afsd arguments from the
config canonical.