afsrobot: support local hostname in the configuration
Instead of hardcoding 'localhost' in the configuration file section name
indicate the local system, allow for the section to be named after the actual
hostname, short or fully qualified.  For example, on host "mantis",

   [host:mantis]
   ... options for the (local) host 'mantis' ...

instead of

   [host:localhost]
   ... options for the local host 'mantis' ...

This has the advantage of being able to share the same configuration file on
the cluster of test hosts. Also, this seems to be more intuitive to new users.

The old naming convention is still supported to avoid breaking old config
files.