afsutil: put the last 20 lines in the CommandFailed exception
Save the last 20 lines of the command output (stdout and stderr)
into the CommandFailed exception to improve error reporting. Do
this even when we are not capturing the full output.

Fix the chomp when logging output lines.

Fix the newline char used to join the output lines to report errors.