support per-user package installs
Get pip the --user option when the install script is invoked as a non-root user
to support per-user package installs.  When invoked as root, do system-wide
package installs.

When doing the system-wide installs (as root), build the dist tarball as the
regular user (that is the same user that owns the package source directory), so
the files generated are not owned by root and can be cleaned.

Add uninstall scripts to undo the package install.