afsrobot: put keytabs in the userdir instead of /tmp
Change the configure defaults to put keytabs in the user's data
directory instead of the global /tmp directory, where they clobber
other user's keytabs.

[It would be nice to use interpolation to set the path to the user's
data dir, which is in the [paths] section. However, interpolation of
data from a different section is not supported until the 3.4 version of
the ConfigParser, and I don't feel like backporting that feature.]