readme: document sudo setup and mention /etc/hosts
Document how to setup sudo and mention we can't have loopback addresses
for the hostname in /etc/hosts.

Move the build instructions before the test instructions, since we need
binaries to test.

Remove the instructions to set the dest option, since this will be found
automatically if we run setup in the top level source directory.