solaris: avoid magic .:mount path to mount root.cell
The special .:mount mount point is broken on Solaris when -dynroot is
given without -fakestat. Avoid using it when mounting the root.cell
volume and setting the acls in root.afs and root.cell root vnodes.

Instead of using .:mount, create a mount point called /afs/.<cell>/.afs
which is a read-write mount point to the root.afs volume.