solaris 10: fix ps usage on solaris 10
Solaris 10 has two different ps programs. The one under /usr/bin takes the
Sys-V style options (e.g. ps -ef), and the one under /usr/ucb/bin takes the BSD
style options (e.g. ps ax).  We can get one or the other depending on the order
of the directories in the PATH.  The /usr/ucb/bin path is gone on solaris 11
and the new ps program takes both styles.

So, when running on solaris, be sure to avoid the ps in /usr/ucb/bin and always
use one ps style.

I've not tried this on FBSD yet, so we may need to change it again when FBSD
support is added.