build: add --srcdir option for out-of-tree builds
Add the afsutil build --srcdir option to support out-of-source-tree
builds.  You may want to do an out-of-source-tree build to avoid
littering your source tree with build artifacts or to build with your
source tree in /afs but write the build artifacts to a fast local
filesystem.

Example usage:

    $ mkdir /tmp/mybuild
    $ cd /tmp/mybuild
    $ afsutil build --srcdir /afs/mycell/myfiles/openafs

The resulting object and executable files will be written to the current
directory.  The configure script will be generated in the --srcdir if it
does not already exist.