install: add symlinks to scripts on old solaris
Create symlinks to the scripts installed in the non-stardard location on old
solaris so we have the commands in our PATH (as root, non-interacive) without
messing with profiles.

The pip installer embeds a shebang in our scripts with the correct path
to the python we want to use, so just a symlink will do.