teardown: carp and carry on if we cant remove the keytab
It seems some versions of solaris have an rm that have old semantic for
the -f option, so just warn and ignore if the keytab removal fails.

While we are here, remove the local path check for the keytab, since
this could be a remote machine.

Also fix an old log message that still had a trailing new line.