install: patch libtool on solaris 11.3
Solaris 11.3 introduced a change in the nm utility which breaks libtool. The
output of nm -p was fixed to print the correct symbol type for uninitialized
common variables. This breaks the libtool symbol table detection. The bug has
been reported [1] to the libtool project, but a fix has not been released yet
(as of August 2016)

[1]: http://lists.gnu.org/archive/html/bug-libtool/2016-01/msg00000.html