Disable annoying pip version check messages
Starting with pip version 6.0, pip will phone home and print messages to
stderr if a newer version is a available on pypi.  This is annoying
since we normally used a packaged pip, which is always older than the
pypi version.  This also makes the install take longer since pip has
to contact pypi for the version check.

Set the PIP_DISABLE_PIP_VERSION_CHECK env var to disable this check on
versions of pip which do the version check. This is easier than trying
to detect if the pip installed has the new --disable-pip-version-check
command line option.