Add requires-gfind tag to tests
The find -noleaf option used in the find tests is intentional and is
specific to gnu/find.

Add the require-gfind to test which require gnu/find so those tests can
be exluded when gnu/find is not available, e.g., old versions of solaris
when gnu tools have not been installed.

Use the ${GFIND} variable to run gnu/find, since the name of the find
executable varies by system, e.g. find on linux, gfind on solaris.