fix hang during client.acl test
The rm command on solaris will prompt the user, hanging the
test. This is not a shell script. Please use keywords.

The tests were not testing the delete rights, since the user didn't have
list rights on the directory. Add the list rights for the duration of
the delete tests.