afsutil: check for bosserver error on startup
Bail and return an error in the transarc mode server init script when
the bosserver fails to start.  This will cause 'afsutil start server' to
fail with an error when the bosserver fails to start.

Note: Currently the bosserver will exit with 0 when started with an
unrecognized option, so this change to the init script is only useful
for other errors until the bosserver is fixed.