afsrobot: improve setup and teardown code
Add a new data class to clean up the setup and teardown code, which was
excessively verbose.

Prepare the lists of hosts for each type of setup and teardown stages
before executing them, and use those lists to do the setup and teardowns
by stage for each set of hosts for that stage.  This is closer to the
original design, but hopefully more maintainable.