use afsutil to remove fake keytab files in teardown
Instead of running the rm command, which may not be available under sudo
for the current user (and may prompt for conformation, blocking
afsrobot), add a new afsutil subcommand to remove the file.

This commit changes the afsutil command line interface:

* `afsutil fakekey` is replaced with `afsutil keytab create`
* `afsutil setkey` is replaced with `afsutil keytab setkey`
* `afsutil keytab destroy` has been added to remove keytabs

The `afsutil keytab destroy` will only remove files which appear to be
valid keytab files.  The --force flag is used to ignore errors during
teardown in the event the keytab file is already missing.