ssh: support local host as well
In order to simplify some test cases, allow the ssh commands to work on
the local host as well.

In addition, add a new --include <hosts> option to 'ssh exec' so that
commands may be sent to a single host without using exclude list tricks
or providing a special config file.  For compatibility, the default host
list remains config.opthostnames when --include is not specified.