cell.py: support remote DB hosts
If afsrobot setup (afsutil newcell) is run from a host that is not the
"primary" DB server, newcell vos and pts commands will be delayed.

To explain why, let's define a few terms:
primary_db	the first “isdbserver = yes” host in the config file
primary_fs	the first “isfileserver = yes” host in the config file
control_host	the host where you run afsrobot teardown/setup/test

When running newcell (from the "control_host"), we adjust the primary_db
CellServDB to reference only itself.  However, if the control_host is
not the same as the primary_db, the control_host CellServDB will still
contain all the DB servers, not just the primary_db.  This can cause
significant delays as newcell vos and pts commands attempt to contact DB
servers that are not running yet.

Modify newcell to restrict the control_host CSDB as well (and restore it
when the rest of the servers are added).