split newcell into newcell/mtroot
Split the Cell.newcell method into two functions; one called newcell()
to setup the new cell on a server, and a new function called mtroot(),
to setup the root volumes in the afs namespace on a host running the
client.

Remove the old --no-client mode from the afsutil new cell cli subcommand
and update the afsrobot runner to call mtroot after newcell. (For now,
run mtroot on the localhost; but in the future mtroot could be invoked
on a specified host running a client.)