afsrobot: do not assume the localhost is a server and/or client
Fix afsrobot setup when the localhost is not a server and/or client, in
multi-host configurations.

Instead, run afsutil newcell on the first server host (by config order),
and run afsutil mtroot on the first client host (by config order).

Later commits may add a config option to specify which host to use,
regardless of order in the config file.