afsutil: workaround ubik first write bug for ptserver too
Due to the ubik bugs in 1.6.18+, the first write to the ptserver may
fail and quorum is dropped. The work around is to wait for more than 75
seconds and retry.  This work around was previously added to the first
vos commands. This comment adds the ubik bug workaround to the first pts
create command.