afsutil: shutdown servers before changing cell info
Do not assume the server processes are not running when newcell is
invoked. If newcell is retried, or if someone manually ran bos commands,
the database and/or the fileservers could already be running, and the
BosConfigs may not be empty.

Since we start the database servers in stages, starting the first db
server, then adding the other hosts, be sure all of the servers are
shutdown first, and restart them only after the cell info has been
configured.

Note: Since we restart existing bnodes, the command line options are
currently not updated.