afsutil: wait for the first ptserver to initialize
There is a race between when the first ptserver is started and the first
time we try to create an entry. Be sure the ptserver is started by
giving it some time to start and checking the state with udebug before
preceeding.

(cherry picked from commit b46fe68b83240b01f435ad9321da76ffd0e2ad3f)