afsutil: rename primary servers to first servers
The first db hostname specified is used to create the first database
server and then the other servers are added. We called this the primary
server, but this is a misnomer, because the db servers are really peers.
Instead, this is the first server in the list, so just use the first
element of the list instead of a separate variable.

Do the same or the fileserver as well. Remove the "primary" fileserver
and use the first element of the list of fileservers to house the
read-write root volumes.