fix file name inconsistencies
The names of the various files are a mess and wildly inconsistent.
Rename the files, dirs, and env vars to be more consistent. The name of
the overall project is openafs-robotest, but the frontend tool is now
called afsrobot.  Rename all the various files and directories
introduced over time to be consistent with the name afsrobot.

The user data directory is no longer hidden. The new default path is
$HOME/afsrobot. The new name of the afsrobot configuration file is
$HOME/afsrobot/afsrobot.ini

The default test cell name is still named 'robotest'.