afsutil: add sh() dryrun option
Add option to sh() to just print the command line which would be run to
stdout and return without executing the command.

Also update the logging prefix handling to print the prefix when running
the command and print the running message to the log at debug level even
when quiet is set.