afsrobot: support host specific server options
Merge the 'options' section with the 'cell' section of the config
file. These are the common command line options for all the hosts
in the cell.

Support overriding the command line options in each separate host
section, so hosts can have differing command line options.

Remove the 'dafs' option, since afsutil newcell will determine if dafs
is in effect based on the fileserver programs names given.