remove install.sh
Use the setup.py 'install_requires' properly and deep six the install
and uninstall shell scripts. Remove the pip commands run in the
preinstall scripts.

Support three install modes:

 * sudo make install --> requires root and installs to PREFIX
 * make install-user --> installs to $HOME/.local
 * make install-dev  --> development mode

Save the PREFIX in a file called Makefile.config file for the uninstall.
This file is generated and sets the PREFIX to a default value of
/usr/local, or /opt when /usr/local is not present.

To install to a custom prefix:

   sudo make install PREFIX=/some/other/path