test: avoid creating a rogue volume during create
The user should not be able to create a new volume if another volume
with the same id already exist. In order to test if this is true, this
commit introduces a test that checks if 'vos create' fails when another
volume (orphan) with the same id already exist in another partition.