test: avoid creating a rogue volume during release
The user should not be able to release a RO volume if another volume
with the same id already exist. In order to test if this is true, this
commit introduces a test that checks if 'vos release' fails when another
volume (orphan) with the same id already exist in another partition.