test: avoid creating a rogue volume during clone
The user should not be able to create a clone with an id that is already
being used by another volume. In order to test if this is true, this
commit introduces a test that checks if 'vos clone -toid‘ fails when
another volume (orphan) with the same id already exist in another
partition.