test: avoid creating a rogue volume during backup
The user should not be able to create a backup with an id that is
already being used by another volume. In order to test if this is true,
this commit introduces a test that checks if 'vos backup‘ fails when
another volume (orphan) with the same id already exist in another
partition.