test: avoid creating a rogue volume during move
The user should not be able to move a volume to a server that already
stores an orphan volume with the same id. In order to test if this is
true, this commit introduces a test that checks if 'vos move‘ fails when
the destination server already has another volume (orphan) with the same
id in a different partition.