afsutil: add misc.py
Create a new module for common utility functions. Deduplicate the
optlists2dict() function.

Use flatten() in lists2dict() to more fully support the command line
interface.  That is, argparse converts the flag "-o a=1 b=2" to ['a=1',
'b2=1'], and the flags "-o a=1 -o b=2" to the list of lists [['a=1'],
['b=2']]. flatten() converts the latter to the former so both forms
are handled uniformly.