afsutil: add ini file support
Rework the afsutil command line handling to be more maintainable
and add ini file support.

A global ini file call /etc/afsutil.ini can be used to set sitewide
defaults for any command line option. The user may overwride sitewide
defaults with a per user file called ~/.afsutil.ini

The ini file may have one section for each afsutil subcommand. Each
section may have an option with the same name as the command line
option.

Complex options are supported in a subsection. For example, the newcell
command supports multiple --options flags to specify options for
clients and servers. This are supported in the ini file by specifying
a section called [newcell.options] and giving a value for each --option
that would be given on the command line.

Example ini file:

    [install]
    chdir = /home/mmeffie/src/openafs

    [newcell]
    cell = robotest
    admin = robotest.admin
    realm = ROBOTEST
    fs = mantis deb9a
    db = mantis

    [newcell.options]
    bosserver =
    dafileserver =
    davolserver =
    mantis.dafileserver = -d 1 -L
    mantis.davolserver = -d 1

The old scripts directory and the largish afsutil extensionless python
file (which was difficult to debug since it could not be imported) has
been replaced with the regular module file __main__.py and a set of
argparse wrappers in cli.py. The argparse wrappers provide a declarative
way of specifying the commmand line subcommands.

The setup has been changed to use the new style entry points. The
afsutil commands can now be invoked in the usual way as 'afsutil' or
with the python -m option, e.g.,

   python -m afsutil <command> [options]

This commit changes the 'afsutil keytab <command>' to 'afsutil
kt<command>', which is not compatible with previous versions, but
simplifies the code and make the ini file more straightforword.

Remove afsutil sub-commands in afsrobot runner.py, which makes this code
cleaner too.