afsutil: detect kernel devel versions if given a empty list
Change check from 'if x is None' to the more pythonic 'if not x' to also
handle the case of x being an empty list. There are many other checks
like this which should probably be updated.

This change fixes afsutil package, which fails to detect available
kernel-devel versions because a empty list was passed to a function.