install: remove preinstall scripts
These are better done via packaging (e.g. rpm, debs) or with cm systems
(e.g. ansible, cloud-init).  Much of this preinstall was to get pip and
distutils installed on the system. Robotframework actually does not
depend on these, and we should remove our dependencies as well, which
would basically make this obsolete.

Another point is, often we need to install on new systems, and that just
leads to disappointment, so just remove these scripts for now, and work
on removing our non-core dependencies.

This breaks the 'bootstrap.sh' script for now.