Do not recursively mount root.cell in the mount point tests
We currently reuse the root.cell in the mount point tests just because it
is a known volume, but this cause the root.cell to be recursively
mounted, which is out of scope for the simple mount point tests.  Avoid
that situation by taking the time to create a new volume for the
mount point tests.