Desupport dev web server background mode
Always run the basic web server for showing test reports in the
foreground.  This minimal is meant as a quick way to view a test logs
during test development and initial, quick use.

Running in the background has been shown to be unreliable and overly
complicated due to issues like stale pid files or occasional
exceptions.

Testers are encouraged to setup the document root for a proper httpd
server, with proper systemd configuration if the server is needed as a
service.

This change has the nice side effect removing code, simplifying the
afsrobot cli, and removing a configuration option.