Specify file server when setting up test volumes
The 'Create Volume' keyword defaults to the current host, which fails if
the control host is not a fileserver. Fix this by specifying the
fileserver when creating test volumes. The test cases pick the first
host in the list of AFS_FILESERVERS, which is the list of fileservers in
the cell.

Also use test variables to specify the test volume partitions to use.