Avoid python2 print statement in the pag test cases
If the interpreter is python3 (which is should be at this point),
the print statement is a sytax error. Just use sys.stdout.write()
instead of a print statement to support py2 and py3.

[Don't bother with the from __future__ import print_function in this
oneliner, but just use the sys.stdout.write function which is python
2/3 agnositic.]

Thanks to Cheyenne Wills for pointing out the python2-ism here.