Wait for instances in prepare
Remote instances may not be reachable for a few moments after they were
created. Add a wait task in the prepare stage to wait until we can ssh
to them.  This should actually be done in 'create', but it easier to do
this in 'prepare' because we have an inventory after 'create' is done.