Avoid molecule cleanup before destroy in build scenarios
The cleanup playbook is intented to be used to cleanup external
resources allocated during prepare. We are abusing cleanup in the build
scenarios as a convienence to explicitly clean the build directory.

Update the build scenarios to avoid the cleanup before destroy, since
the instances may not exist, in which case cleanup fails and interrupts
destroy.