Convert molecule build scenarios to drive independent templates
Add a set of templated molecule scenarios to build OpenAFS binary and
source distribution archives.

'afs-scenario setup' can be used to create the driver specification
(driver.yml) to specify the molecule driver to be used to create guests.

Add the make builds target as a help to build for different targets and
versions, using molecule environment files to specify the images and
versions.

Remove the old build playbooks.