Refactor molecule build scenario
Move the build tasks to a shared directory and consolidate build
scenarios into a single 'build' scenario with the afs_build variable
specifying the build target (packages, bdist, or sdist).  Move the code
checkout to a common file as well.

For bdist builds, optionally read the configure options from a user
provided file, otherwise use default values based on the afs_bdist_style
variable. Add the afs_bdist_name variable to override the bdist
component of the generated installation archive filename in order to
support different configure options for a given software version.

With this change, we have less clutter in the molecule list output, as
well support for different build targets in parallel on different hosts
in a given scenario (if one updates the molecule file to add more
hosts).