Use keytab instead of password for logins
Convert tests to use a keytab instead of a password string to get
tokens.  By default, the admin user created by the openafs roles is
'<username>.admin', and the regular user is '<username>' with a pts id
that matches <username>'s uid.

Requires robotframework-openafslibrary v0.7.2, which added the keytab
argument to the Login() keyword.