Rewrite molecule scenario cookiecutter template
Move the molecule playbook files out of the "default" directory so they
can be shared if sub-scenarios are created under ./molecule.

Move the tasks for general test instances setup to the new
'prologue.yml' playbook.

Add the requirements.txt file to make it easier to install ansible and
molecule via pip.

By default, let molecule automatically install the OpenAFS Ansible
Collection with ansible-galaxy.  This can be disabled by setting the
'enabled' key to False in the dependency section.  The source and
version of the collection can be set in the collections.yml requirement
file.

Add the build.yml playbook to automate the building of packages, binary
and source distributions after the prologue and before deploying the
realm and cell. Add the 'external_build' cookiecutter parameter to
generate a scenario without build playbooks.