Add add. checks for creating mountpoint in root.cell
In older releases of AFS there seems to be a bug when
removing mountpoint via fs rmmount.
The mountpoint seems to be deleted but with a stat PATH
it can be checked that the mountpoint still exists.
This could be a cachemanager problem, fs flusha can fix it.
In newer releases (1.8.x) the problem seems to be fixed
already.
It could also be that this wrong behaviour only occurs if
the PATH of the mountpoint is in root.cell volume.

This additional robo test will report the problem.