Support 'Compare Files' keyword on Solaris
Unfortunately, the 'Compare Files' keyword invokes the diff command with
a GNU specific option.

In the future, we should make a python based keyword to do the
comparison, but for now use GNU diff when running on Solaris to work
around this issue.

This allows the tests to run on Solaris when the requires-multi-fs is
option is enabled.

[mmeffie: squashed commits, updated message]