Convert the robot files to pipe separated format
The space separated format tends to be error prone, so convert the robot
files to the pipe separated format style.

The conversion was done by first converting the robot files to tab
separated format with the robotidy utility, then running our
tools/robotab script to convert the tables to pipe separated and column
aligned.

    $ robotidy --separator tab tests/
    $ tools/robotab tests/