cookiecutter: Update lint rules for ansible-lint 6
ansible-lint 6 added a complaint about builtin tasks. Add a skip rule
to accept those for now.

Seems ansible-lint was dropped from molecule[ansible] bundle, so add
ansible-lint to the pip rerequirements to be sure it is installed in our
virtualenv for testing the cookiecutter template.

Change `make test` to be very verbose. Use `make check` for quieter
mode test.