cookiecutter: Update quickstart
Move the driver section back to the molecule file so we can remove the
dependency on git in the quick-start and remove the hidden base config
file.

Limit the collection source to galaxy or git to simplify the template
and collection.yml

Update the instructions to install cookiecutter as a prereq, like
virtualenv.

Remove some unused items in the cookiecutter json file.