cookiecutter: Prompt for driver again
Reintroduce a driver selection in the cookiecutter prompts.  This should
make it easier to create scenarios for prebuilt guests, which is not too
uncommon for development.

The current version of cookiecutter is very limited and does not support
condtional questions, so instead of asking for the driver name, list
present a flat list of driver types/subtypes.