cookiecutter: Add set_hostname install prepare option for proxmox images
Add a feature to set the instance hostname during the prepare stage.
This is needed for the proxmox driver.

Also, add the proxmox driver settings in the template, but do so as
env vars.  It would be better to prompt for those, but cookiecutter
does not support condional prompts yet, so for now just let the user
specify them as env vars or edit the generated molecule file.